requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
PyMuPDF
pandas
faiss-cpu
openai
python-dotenv
requests